requirements.txt
yara-python
requests
reportlab
# Core dependencies
pyyaml
# No extra Python packages needed for nuclei
tk
